requirements.txt
# Dive AI V29.4 Requirements

# Core Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1

# HTTP & WebSocket
aiohttp>=3.9.1
websockets>=12.0
requests>=2.31.0

# Desktop Automation (for channels)
pyautogui>=0.9.54
pytesseract>=0.3.10
Pillow>=10.2.0
opencv-python>=4.9.0

# Desktop IDE
PyQt6>=6.10.0
PyQt6-WebEngine>=6.10.0

# System Tray
pystray>=0.19.5

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6
